Merge Results Page with Address Lookup Tool Integration (#23)
* Results and ALT pages now call correctly.

Before, when the search candidates function was called, it would always point to either end of the if/else block. Now it doesn't. This is great.

* Party colours now display properly.

* Candidates are now randomized in Address Lookup.

* Get the ALT styling back.

The styling got eaten somewhere along the way. It is now restored.

* Pages are now in the v2 folder instead of outside.

The files can be copied at the end into v1 if necessary.

* Caved and just used GET for all results.

AJAX was a mistake.